# docs/openapi/frontend-surface.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FERN Frontend Surface",
    "version": "2026-03-31",
    "description": "Source-derived OpenAPI contract for the current frontend-facing gateway surface. It includes endpoints actively used by the frontend app or intentionally exposed for frontend diagnostics. It excludes internal control-plane routes, but now includes the auth session lifecycle, shared paged staff list envelopes with total-count metadata, the stock-safe sales, payroll, and procurement actions already exercised by the frontend, the dedicated cashier POS workspace at exact /pos built on existing protected sales and product endpoints, the protected non-cashier staff order queue at exact /order, and the public customer ordering plus public order-status flow under /api/v1/sales/public."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8180",
      "description": "Local gateway"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/reports/sales": {
      "get": {
        "summary": "Sales summary",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales summary list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesSummary"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/reports/expenses": {
      "get": {
        "summary": "Expense summary",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense summary list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseSummary"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/reports/inventory-movements": {
      "get": {
        "summary": "Inventory movement summary",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movement summary list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryMovementSummary"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/reports/low-stock": {
      "get": {
        "summary": "Low stock snapshot",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Low-stock rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LowStockSnapshot"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to access the requested scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorPayload": {
        "type": "object",
        "required": [
          "timestamp",
          "error",
          "message"
        ],
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        }
      },
...
      "SalesSummary": {
        "type": "object",
        "required": [
          "outletId",
          "businessDate",
          "saleCount",
          "subtotal",
          "discount",
          "taxAmount",
          "totalAmount"
        ],
        "properties": {
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "businessDate": {
            "type": "string"
          },
          "saleCount": {
            "type": "integer"
          },
          "subtotal": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "taxAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          }
        }
      },
      "ExpenseSummary": {
        "type": "object",
        "required": [
          "outletId",
          "businessDate",
          "sourceType",
          "expenseCount",
          "totalAmount"
        ],
        "properties": {
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "businessDate": {
            "type": "string"
          },
          "sourceType": {
            "type": "string"
          },
          "expenseCount": {
            "type": "integer"
          },
          "totalAmount": {
            "type": "number"
          }
        }
      },
      "InventoryMovementSummary": {
        "type": "object",
        "required": [
          "outletId",
          "itemId",
          "businessDate",
          "txnType",
          "netQuantityChange"
        ],
        "properties": {
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "businessDate": {
            "type": "string"
          },
          "txnType": {
            "type": "string"
          },
          "netQuantityChange": {
            "type": "number"
          }
        }
      },
      "LowStockSnapshot": {
        "type": "object",
        "required": [
          "outletId",
          "itemId",
          "itemCode",
          "itemName",
          "qtyOnHand",
          "minStockLevel"
        ],
        "properties": {
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemCode": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "qtyOnHand": {
            "type": "number"
          },
          "minStockLevel": {
            "type": "number"
          }
        }
      }
    }
  }
}